reduce arm double usage remove unnecessary float cast

+++ src/layer/arm/gemm_arm_asimdhp.cpp
@@ -2282,7 +2282,7 @@ static void get_optimal_tile_mnk_fp16sa(int M, int N, int K, int constant_TILE_M
 {
     // resolve optimal tile size from cache size
     size_t l2_cache_size = get_cpu_level2_cache_size();
-    int tile_size = (int)sqrt((float)l2_cache_size / 3 / sizeof(__fp16));
+    int tile_size = (int)sqrtf((float)l2_cache_size / 3 / sizeof(__fp16));
 
     TILE_M = std::max(8, tile_size / 8 * 8);
     TILE_N = std::max(4, tile_size / 4 * 4);
