+++ src/layer/arm/mish_arm.cpp
@@ -109,7 +109,7 @@ int Mish_arm::forward_inplace(Mat& bottom_top_blob, const Option& opt) const
 #endif // __ARM_NEON
         for (; remain > 0; remain--)
         {
-            *ptr = *ptr * tanh(log(exp(*ptr) + 1.f));
+            *ptr = *ptr * tanhf(logf(expf(*ptr) + 1.f));
             ptr++;
         }
     }
@@ -171,7 +171,7 @@ int Mish_arm::forward_inplace_bf16s(Mat& bottom_top_blob, const Option& opt) con
         for (; remain > 0; remain--)
         {
             float v = bfloat16_to_float32(*ptr);
-            v = v * tanh(log(exp(v) + 1.f));
+            v = v * tanhf(logf(expf(v) + 1.f));
             *ptr = float32_to_bfloat16(v);
             ptr++;
         }
