+++ src/layer/tanh.cpp
@@ -38,7 +38,7 @@ int TanH::forward_inplace(Mat& bottom_top_blob, const Option& opt) const
 
         for (int i = 0; i < size; i++)
         {
-            ptr[i] = static_cast<float>(tanh(ptr[i]));
+            ptr[i] = tanhf(ptr[i]);
         }
     }
